win_hwd: Create common context for different collectors
