+++ CMakeLists.txt
@@ -1603,6 +1603,7 @@ set(INTERNAL_COLLECTORS_FILES
         src/collectors/common-contexts/mem-swap.h
         src/collectors/common-contexts/mem-pgfaults.h
         src/collectors/common-contexts/mem-available.h
+        src/collectors/common-contexts/power-supply.h
 )
 
 set(INTERCOMMUNICATION_COLLECTORS_FILES
