+++ src/collectors/common-contexts/common-contexts.h
@@ -35,5 +35,6 @@ typedef void (*instance_labels_cb_t)(RRDSET *st, void *data);
 #include "disk-await.h"
 #include "disk-svctm.h"
 #include "disk-avgsz.h"
+#include "power-supply.h"
 
 #endif //NETDATA_COMMON_CONTEXTS_H
